fixed another bug.
git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@4364 a326200e-df2c-42e0-81fa-597e909af41d

+++ x10.lib/src/types.h
@@ -1,7 +1,7 @@
 /*
  * (c) Copyright IBM Corporation 2007
  *
- * $Id: types.h,v 1.14 2007-08-21 06:10:25 ganeshvb Exp $
+ * $Id: types.h,v 1.15 2007-08-21 11:26:33 ganeshvb Exp $
  * This file is part of X10 Runtime System.
  */
 
@@ -13,7 +13,7 @@
 #include <sys/types.h>
 
 /* async handler */
-typedef long long x10_async_arg_t;
+typedef long x10_async_arg_t;
 typedef int x10_async_handler_t;
 typedef int x10_place_t;
